Rename AsanaHelper to AsanaAPI.
While it's still a helper, the AsanaFacade class (coming in the next commit) is
also a helper.  This is a bit more explicit, especially because the AsanaAPI
class deals *only* with API stuff and should have no additional interaction
with the main Asana application.  At some point in the future, this will likely
be in a separate package.