Switch to addProjects() from setProjects()
Instead of overwriting the entire projects table, we now must merge in separate
projects.  This is because commit 869f903 changes to querying for projects by
workspace ID, not all at once.  The change in this commit also includes
duplicate project protection.  It's currently name based, but should be project
ID based soon.