Load workspaces and projects in background
There's still a few issues with this - the initial attempt at loading
(immediately after adding API key) fails with invalid authentication.  The
second attempt is successful, but the ExpandableListView doesn't update,
despite manually calling onContentChanged().  Relaunching the app a third time
results in a populated ExpandableListView, but the user must manually cancel
the loading process to access it.  This, obviously, is far from ideal.