Initial attempt to use the nav bar for Workspaces
This places the associated projects in the main pane, just like on the mobile
site, but it's broken.  Why?  My guess is because the workspace cursor is still
empty by the time execution gets to creating a project cursor.  This leads to
an uncaught exception.  The solution is probably to test the count of
workspaceCursor before defining projectCursor, or to use onDataSetChanged() and
some form of notification.  Uris support notifications, so that's probably the
safest way to go.